requirements.txt
openai>=1.12.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
pydantic>=2.5.0
tenacity>=8.2.3
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdown>=3.5.0
pyyaml>=6.0.0
